requirements.txt
Django==4.1.3
django-cleanup==6.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.3.0
PyMySQL==1.0.2
django-rest-passwordreset == 1.3.0